+++ pkgs/intl4x/lib/calendar.dart
@@ -1,0 +1,13 @@
+// Copyright (c) 2026, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+/// Calendar types and weekday information.
+///
+/// Provides [Calendar] and [Weekday] enums for working with calendar systems
+/// and week-start information.
+/// @docImport 'src/options.dart';
+library;
+
+export 'src/locale/locale.dart' show Locale;
+export 'src/options.dart' show Calendar, Weekday;