+++ pkgs/intl4x/lib/src/first_day_of_week/first_day_of_week_stub_4x.dart
@@ -1,0 +1,9 @@
+// Copyright (c) 2026, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import '../locale/locale.dart';
+import '../options.dart';
+
+Weekday getFirstDayOfWeek4X(Locale locale) =>
+    throw UnimplementedError('Cannot use ICU4X in web environments.');