+++ pkgs/intl4x/lib/src/plural_rules/plural_rules_impl.dart
@@ -4,7 +4,6 @@
 
 import '../locale/locale.dart';
 import '../utils.dart';
-import 'plural_rules.dart';
 import 'plural_rules_options.dart';
 import 'plural_rules_stub.dart'
     if (dart.library.js_interop) 'plural_rules_ecma.dart';
@@ -21,3 +20,6 @@ abstract class PluralRulesImpl {
   static PluralRulesImpl build(Locale locales, PluralRulesOptions options) =>
       buildFormatter(locales, options, getPluralSelectECMA, getPluralSelect4X);
 }
+
+/// Defines the locale-specific grammatical categories for plural forms.
+enum PluralCategory { zero, one, two, few, many, other }